# programming_server_crossplatform_applications/labs/lab_23/doc/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "lab12",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/TS": {
      "get": {
        "tags": [
          "default"
        ],
        "summary": "getTs",
        "responses": {
          "200": {
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "post": {
        "tags": [
          "default"
        ],
        "summary": "postTS",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example":
                {
                  "id": 2,
                  "name": "Alexandra",
                  "phone": "22222222"
                }

              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": { }
            }
          }
        }
      },
      "put": {
        "tags": [
          "default"
        ],
        "summary": "putTS",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "example":
                {
                  "id": 2,
                  "name": "hhhh",
                  "phone": "22222222"
                }

              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      },
      "delete": {
        "tags": [
          "default"
        ],
        "summary": "deleteTS",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK message for delete",
            "content": {
              "application/json": { }
            }
          }
        }
      }
    }
  }
}